+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -3,7 +3,7 @@ import { ATMB202305 } from './20230502_ATMB';
 import { Cannes2024 } from './20240101_Cannes';
 import { Cotentin2023 } from './20230101_Cotentin';
 import { GrandChatellerault2024 } from './20240201_GrandChatellerault';
-import { GrandPoitiers } from './GrandPoitiers';
+import { GrandPoitiers } from './20230927_GrandPoitiers';
 import { Idfm } from './20210520_IDFM';
 import { LaRochelle20232024 } from './LaRochelle20232024';
 import { Lannion } from './Lannion';
